Extract reusable Faq component and landing FAQ data
Made-with: Cursor
